requirements.txt
streamlit
ibm_db_dbi==3.0.3
mysql-connector-python==8.0.23
cx_Oracle==8.3.0
pandas==1.3.5
psutil